backend/requirements.txt
openai-whisper
pytest
pytest-mock
librosa
pyin
